Made minor parameter-based optimizations
* Added const reference to shared_ptr and QString parameters.
* Changed a for loop from int index to size_t index when passing index into an array's [] operator.

+++ SIMPLVtkLib/Visualization/Controllers/VSFilterViewSettings.cpp
@@ -1227,7 +1227,7 @@ bool VSFilterViewSettings::isFlatImage()
 
   // Check dimensions
   int* dims = imageData->GetDimensions();
-  for(int i = 0; i < 3; i++)
+  for(size_t i = 0; i < 3; i++)
   {
     if(dims[i] <= 1)
     {
